Productos & cupones de descuentos módulos reformulados hechos / api resend con algunas funciones

+++ DOCUMENTS/Database/Procedimientos/cupones_descuentos.sql
@@ -42,26 +42,27 @@ SELECT * FROM cupones_descuento;
 CREATE PROCEDURE spu_eliminar_cupon_descuento
 (
     p_id_cupon_desc INT,
-    p_id_tienda INT,
-    p_aplica_producto BOOLEAN
+    p_id_tienda INT
 )
 LANGUAGE plpgsql
 AS $$
 BEGIN
-    CASE p_aplica_producto
-        WHEN FALSE THEN
-            DELETE FROM cupones_descuento
-            WHERE id_cupon_desc = p_id_cupon_desc
-            AND id_tienda = p_id_tienda;
+    IF EXISTS(
+        SELECT 1 
+        FROM  cupones_descuentos_productos
+        WHERE id_cupon_desc = p_id_cupon_desc
+    ) THEN
+        DELETE FROM cupones_descuentos_productos
+        WHERE id_cupon_desc = p_id_cupon_desc;
 
-        WHEN TRUE THEN
-            DELETE FROM cupones_descuentos_productos
-            WHERE id_cupon_desc = p_id_cupon_desc;
-
-            DELETE FROM cupones_descuento
-            WHERE id_cupon_desc = p_id_cupon_desc
-            AND id_tienda = p_id_tienda;
-    END CASE;
+        DELETE FROM cupones_descuento
+        WHERE id_cupon_desc = p_id_cupon_desc
+        AND id_tienda = p_id_tienda;
+    ELSE
+        DELETE FROM cupones_descuento
+        WHERE id_cupon_desc = p_id_cupon_desc
+        AND id_tienda = p_id_tienda;
+    END IF;
 END;
 $$;
 
@@ -75,10 +76,13 @@ CREATE PROCEDURE spu_modificar_cupon_descuento(
     p_fecha_expiracion DATE,
     p_usos_maximos INT,
     p_usos_actuales INT,
-    p_aplica_producto BOOLEAN
+    p_aplica_producto BOOLEAN,
+    p_productos_aplicados INT[]
 )
 LANGUAGE plpgsql
 AS $$
+DECLARE
+    p_id_producto INT;
 BEGIN
     UPDATE cupones_descuento
     SET
@@ -90,5 +94,22 @@ BEGIN
     usos_actuales = COALESCE(p_usos_actuales, usos_actuales)
     WHERE id_cupon_desc = p_id_cupon_desc
     AND id_tienda = p_id_tienda;
+
+    IF p_aplica_producto IS TRUE AND array_length(p_productos_aplicados, 1) > 0 THEN
+        DELETE FROM cupones_descuentos_productos
+        WHERE id_cupon_desc = p_id_cupon_desc;
+        FOREACH p_id_producto IN ARRAY p_productos_aplicados
+        LOOP
+            INSERT INTO cupones_descuentos_productos (id_cupon_desc, id_producto)
+            VALUES (p_id_cupon_desc, p_id_producto);
+        END LOOP;
+    ELSIF p_aplica_producto IS FALSE THEN
+        DELETE FROM cupones_descuentos_productos
+        WHERE id_cupon_desc = p_id_cupon_desc;
+    END IF;
 END;
 $$;
+
+SELECT * FROM cupones_descuento
+SELECT * FROM cupones_descuentos_productos
+SELECT * FROM productos
